requirements.txt.txt
opencv-python
numpy
pillow
pygame
dlib
face-recognition
deepface
scipy
flask
